+++ src/account/HISTORY.rst
@@ -5,7 +5,7 @@ Release History
 
 0.2.2
 +++++
-* Update SDK
+* Update SDK `azure-mgmt-subscription` version to 3.0.0 and api-version to 2021-10-01.
 
 0.2.1
 +++++